refactoring of GrantRole component
